+++ allaboutmorphe/MorpheApp/morphe-patches-library/extension-library/src/main/java/app/morphe/extension/shared/ResourceType.java
@@ -43,6 +43,7 @@ public enum ResourceType {
 
     public final String type;
 
+    @SuppressWarnings("unused")
     public static ResourceType fromValue(String value) {
         ResourceType type = VALUE_MAP.get(value);
         if (type == null) {