+++ allaboutmorphe/MorpheApp/morphe-patches-library/extension-library/src/main/java/app/morphe/extension/shared/ResourceUtils.java
@@ -2,7 +2,6 @@ package app.morphe.extension.shared;
 
 import android.app.Activity;
 import android.content.Context;
-import android.content.res.Configuration;
 import android.content.res.Resources;
 import android.graphics.Color;
 import android.graphics.drawable.Drawable;