+++ allaboutmorphe/MorpheApp/morphe-patches-library/extension-library/src/main/java/app/morphe/extension/shared/StringRef.java
@@ -7,6 +7,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
+@SuppressWarnings("ALL")
 public class StringRef {
     /**
      * Simple class to assist with apps that currently cannot use resource patches.