+++ allaboutmorphe/MorpheApp/morphe-patches-library/extension-library/src/main/java/app/morphe/extension/shared/StringTrieSearch.java
@@ -3,6 +3,7 @@ package app.morphe.extension.shared;
 /**
  * Text pattern searching using a prefix tree (trie).
  */
+@SuppressWarnings("unused")
 public final class StringTrieSearch extends TrieSearch<String> {
 
     private static final class StringTrieNode extends TrieNode<String> {