+++ allaboutmorphe/MorpheApp/morphe-patches-library/extension-library/src/main/java/app/morphe/extension/shared/settings/AppLanguage.java
@@ -2,6 +2,7 @@ package app.morphe.extension.shared.settings;
 
 import java.util.Locale;
 
+@SuppressWarnings("unused")
 public enum AppLanguage {
     /**
      * The current app language.