+++ allaboutmorphe/MorpheApp/morphe-patches-library/patch-library/src/main/kotlin/app/morphe/patches/all/misc/fix/changepackageinstaller/ChangePackageInstallerPatch.kt
@@ -14,6 +14,12 @@ import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.util.matchAllMethodIndicesForEach
 import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
+import java.util.logging.Logger
+
+/**
+ * App store source.
+ */
+private const val PACKAGE_INSTALLER_PACKAGE_SOURCE_STORE = 2
 
 /**
  * Spoofs the installer source for all methods/field usage of:
@@ -21,33 +27,58 @@ import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
  * ```
  *  Landroid/content/pm/PackageManager;->getInstallerPackageName(Ljava/lang/String;)Ljava/lang/String;
  *  Landroid/content/pm/InstallSourceInfo;->getInstallingPackageName()Ljava/lang/String;
- *  Landroid/content/pm/InstallSourceInfo;->getOriginatingPackageName()Ljava/lang/String;
  *  Landroid/content/pm/InstallSourceInfo;->getInitiatingPackageName()Ljava/lang/String;
+ *  Landroid/content/pm/InstallSourceInfo;->getPackageSource()I
  * ```
  *
- * @param installerPackageName Installer package name to use. Defaults to the Google Play Store.
+ * @param installerPackageNameProvider Installer package name to use. Defaults to the Google Play Store.
  */
-fun changePackageInstallerPatch(installerPackageName : String = "com.android.vending") = bytecodePatch {
+@Suppress("unused")
+fun changePackageInstallerPatch(
+    installerPackageNameProvider: () -> String = { "com.android.vending" }
+) = bytecodePatch {
     execute {
+        var changesMade = false
+        val installerPackageName = installerPackageNameProvider()
+
         arrayOf(
             "Landroid/content/pm/PackageManager;->getInstallerPackageName(Ljava/lang/String;)Ljava/lang/String;",
             "Landroid/content/pm/InstallSourceInfo;->getInstallingPackageName()Ljava/lang/String;",
-            "Landroid/content/pm/InstallSourceInfo;->getOriginatingPackageName()Ljava/lang/String;",
-            "Landroid/content/pm/InstallSourceInfo;->getInitiatingPackageName()Ljava/lang/String;"
+            "Landroid/content/pm/InstallSourceInfo;->getInitiatingPackageName()Ljava/lang/String;",
+            "Landroid/content/pm/InstallSourceInfo;->getPackageSource()I"
         ).forEach { smali ->
-            methodCall(smali).matchAllMethodIndicesForEach(requireMatches = false) { index ->
+            val methodCall = methodCall(smali)
+            val isIntReturnType = methodCall.returnType == "I"
+            val expectedReturnOpcode = if (isIntReturnType) {
+                Opcode.MOVE_RESULT
+            } else {
+                Opcode.MOVE_RESULT_OBJECT
+            }
+
+            methodCall.matchAllMethodIndicesForEach(requireMatches = false) { index ->
                 val returnIndex = index + 1
                 val instruction = getInstruction(returnIndex)
-                if (instruction.opcode != Opcode.MOVE_RESULT_OBJECT) {
+
+                if (instruction.opcode != expectedReturnOpcode) {
                     return@matchAllMethodIndicesForEach
                 }
 
                 val register = (instruction as OneRegisterInstruction).registerA
-                replaceInstruction(
-                    returnIndex,
+                val smaliInstruction = if (isIntReturnType) {
+                    "const/4 v$register, $PACKAGE_INSTALLER_PACKAGE_SOURCE_STORE"
+                } else {
                     "const-string v$register, \"$installerPackageName\""
-                )
+                }
+                replaceInstruction(returnIndex, smaliInstruction)
+
+                changesMade = true
             }
+        }
+
+        if (!changesMade) {
+            return@execute Logger.getLogger(this::class.java.name).info(
+                "No installer source checks found, no changes applied"
+            )
         }
     }
 }